+++ package.json
@@ -30,6 +30,7 @@
     "@types/isomorphic-fetch": "^0.0.36",
     "@types/leaflet": "^1.9.6",
     "@types/react": "^18.2.22",
+    "@types/svg-path-parser": "^1.1.4",
     "@types/swagger-ui-react": "^4.18.0",
     "@typescript-eslint/eslint-plugin": "^6.7.2",
     "@typescript-eslint/parser": "^6.7.2",
