+++ src/api/operations.ts
@@ -31,86 +31,6 @@ export function moveMedia(
   );
 }
 
-export function getSvgEdit(
-  accessToken: string | null,
-  problemId: number,
-  mediaId: number,
-): Promise<any> {
-  return makeAuthenticatedRequest(
-    accessToken,
-    `/problem?id=${problemId}&showHiddenMedia=true`,
-    null,
-  )
-    .then((data) => data.json())
-    .then((res) => {
-      const m = res.media.filter((x) => x.id == mediaId)[0];
-      const readOnlySvgs: {
-        nr: number;
-        hasAnchor: boolean;
-        path: string;
-        anchors: unknown[];
-        texts: string[];
-      }[] = [];
-      let svgId = 0;
-      let hasAnchor = true;
-      let path = null;
-      let anchors = [];
-      let texts = [];
-      if (m.svgs) {
-        for (const svg of m.svgs) {
-          if (svg.problemId === res.id) {
-            svgId = svg.id;
-            path = svg.path;
-            hasAnchor = svg.hasAnchor;
-            anchors = svg.anchors ? JSON.parse(svg.anchors) : [];
-            texts = svg.texts ? JSON.parse(svg.texts) : [];
-          } else {
-            readOnlySvgs.push({
-              nr: svg.nr,
-              hasAnchor: svg.hasAnchor,
-              path: svg.path,
-              anchors: svg.anchors ? JSON.parse(svg.anchors) : [],
-              texts: svg.texts ? JSON.parse(svg.texts) : [],
-            });
-          }
-        }
-      }
-      return {
-        mediaId: m.id,
-        nr: res.nr,
-        w: m.width,
-        h: m.height,
-        shift: false,
-        svgId: svgId,
-        path: path,
-        anchors: anchors,
-        texts: texts,
-        readOnlySvgs: readOnlySvgs,
-        activePoint: 0,
-        draggedPoint: false,
-        draggedCubic: false,
-        hasAnchor: hasAnchor,
-        areaId: res.areaId,
-        areaName: res.areaName,
-        areaLockedAdmin: res.areaLockedAdmin,
-        areaLockedSuperadmin: res.areaLockedSuperadmin,
-        sectorId: res.sectorId,
-        sectorName: res.sectorName,
-        sectorLockedAdmin: res.sectorLockedAdmin,
-        sectorLockedSuperadmin: res.sectorLockedSuperadmin,
-        id: res.id,
-        name: res.name,
-        grade: res.grade,
-        lockedAdmin: res.lockedAdmin,
-        lockedSuperadmin: res.lockedSuperadmin,
-      };
-    })
-    .catch((error) => {
-      console.warn(error);
-      return null;
-    });
-}
-
 export function downloadUsersTicks(accessToken: string | null) {
   return downloadXlsx(accessToken, `/users/ticks`);
 }
